plugins: adrv9002: improve carrier handling
Handling the carrier frequency is not really straight because it looks
like we have 4 independent controls when in reality they are not, since
the device only has 2 LOs. It all the depends on the LO mappings present in
the current profile. First, the only way a carrier is actually applied (PLL
re-tunes) is if all ports on the same LO, are moved into the calibrated
state before changing it. Not doing so means that we may end up in an
inconsistent state. For instance, consider the following steps in a FDD
profile where RX1=RX2=LO1 (assuming both ports start at 2.4GHz):

 1. Move RX1 carrier to 2.45GHz -> LO1 re-tunes
 2. Move RX1 back to 2.4GHz -> LO1 does not re-tune and our carrier is at
    2.45GHz

With the above steps we are left with both ports, __apparently__, at
2.45GHz but in __reality__ our carrier is at 2.4GHz.

Secondly, when both ports of the same type are at the same LO, which
happens on FDD and TTD (with diversity) profiles, we should move the
carrier of these ports together, because it's just not possible to have
both ports enabled with different carriers (they are on the same LO!). On
TDD profiles, we never move TX/RX ports together even being on the same LO.
The assumption is that we might have time to re-tune between RX and TX
frames. If we don't, we need to manually set TX and RX carriers to the same
value before starting operating...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>