cmake: remove link_directories
The 'link_directories()' should almost never be needed. Moreover, it's not
clear why it was being setted. On top of this 'CMAKE_PREFIX_PATH' is not
the variable that should be used for search for link directories. If one
needs to look for libraries in a different path, typically that means that
all needed libraries will be in that "sysroot". Hence, one can just set
'CMAKE_SYSROOT' from the command line.

Anyways, let's just remove this for now and if the usecase for having this
pops up we can properly add it.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>