plugins: adrv9002: fix profile configuration on windows
We need to add the binary mode to 'fopen()' as it seems that windows treats
text files and binary files differently (on POSIX systems the b mode is
actually ignored). Otherwise, 'fread()' will actually return less bytes
than asked (with the size computed with fseek + ftell) and we will error
out before loading the stream. This fixes it and causes 'fread' to
return the expected number of bytes.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>