plugins: adrv9002: fix carrier handling
There's a problem when handling carrier changes on TDD profiles (where
TX/RX on the same LO are not moved together. Considering that RX1 and TX1
are both on LO1:

 1. TX1 carrier starts at 2.45GHz and primed;
 2. RX1 carrier starts at 2.4GHz and primed;
 3. Change TX1 to rf_enabled and set the carrier to 2.5GHz in the plugin;

With the last steps we end up with 2.4GHz on LO1 which is not expected.
The reason is that we will move RX1 from calibrated (because we moved RX1
to calibrated since all ports on the same LO must in the calibrated state to
change the carrier) to prime after changing TX1 which causes a re-lock on
the RX1 carrier (2.4GHz). Thus, we need to make sure that the port where
we are changing the carrier is always the last one to transition
state. This fixes the issue for TDD profiles and has no effect on FDD
ones.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>