treewide: properly annotate fall-through labels
Need to make sure all fall-through labels are annotated so that clang
does not complain. A macro was added so that it looks better than using
plain compiler __attributes__.

Note that the macros is pretty direct which means there is no check to
make sure the compiler has the proper attribute. Nowadays, it should not
be an issue but if someone complains, we can then handle it at that time.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>