plugins: adrv9002: move combo box to signal blocked API
This is similar to the previous
commit ("plugins: adrv9002: move to widget signal  blocked signal API") but
for all the combo boxes widgets.

On top of that and with
commit 83492f2532d37 ("iio_widget: Add new combo box widget"), now there's
no need to have a custom combo box implementation. Note that the
implementation was also taking care of autonomous updates which is,
essentially, what the block API does. Thus, all conditions are met to drop
the custom combo box implementation.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>