note-for-dev: update note
With the GTK3 based IIO-Oscilloscope, using glade 3.8 is not necessary.
Glade v3.9 and up (preferably use version 3.38) are all suitable for
creating glade files.
If a glade file was created using glade for GTK2, said file needs to be
converted to glade for GTK2, as per the instructions in the note.

Signed-off-by: Cristina Suteu <cristina.suteu@analog.com>