plugins: spectrum_analyser: Fix sampling freq check
When trying to start the spectrum analyzer from the plugin,
any mismatch between the requested sampling rate and the set
rate read back causes the plugin to fail to launch due to
a `return false`.

On many systems, there is a small mismatch where `61440000`
(default rate) will return `61439999`, causing a match failures.
This change simply logs the mistmatch warning, but still allows
the output window to be shown.

Signed-off-by: Kevin Townsend <kevin.townsend@analog.com>