plugins: adrv9002: constrain RSSI reads
With the changes in the device driver API, reading the RSSI on a channel
that's not enabled will trow an error. Hence, instead of flooding the
device with error's (and the kernel log), don't read the RSSI for
channels that are not enabled.

With the above, as we always need to read the ensm widget, make use
iio_widget_update_value() to update it in case port enable is set to pin
(and hence avoiding another call into the device).

Signed-off-by: Nuno Sa <nuno.sa@analog.com>