plugins: adrv9002: allow control RF state in pin mode
There's now the possibility of still controlling the port RF enable
state mode even if the port mode is set to pin. That's possible in case
the driver is in control of the pin controlling the port. Hence, let's
check if the driver supports that at the plugin initialization and act
accordingly. To check it we do the following:

 1) Move the port to pin mode if in spi mod;.
 2) Change the RF state mode to rf_enabled;
 3) If success, then we allow changing the RF state even in pin mode;

Signed-off-by: Nuno Sa <nuno.sa@analog.com>